+++ components/EditScreenInfo.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import React from "react";
 
 import { ExternalLink } from "./ExternalLink";
