+++ components/StyledText.tsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 import { Text, TextProps } from "./Themed";
 
 export function MonoText(props: TextProps) {
