+++ components/Themed.tsx
@@ -3,6 +3,7 @@
  * https://docs.expo.io/guides/color-schemes/
  */
 
+import React from "react";
 import {
   Text as DefaultText,
   useColorScheme,
