+++ package.json
@@ -20,6 +20,7 @@
     "@babel/runtime": "^7.23.5",
     "@expo/metro-config": "^0.10.7",
     "@expo/vector-icons": "^13.0.0",
+    "@react-native-async-storage/async-storage": "^1.21.0",
     "@react-navigation/native": "^6.0.2",
     "@shopify/flash-list": "1.4.3",
     "@shopify/react-native-skia": "0.1.196",
@@ -33,6 +34,7 @@
     "expo-font": "~11.4.0",
     "expo-image-picker": "~14.3.2",
     "expo-linking": "~5.0.2",
+    "expo-media-library": "^15.6.0",
     "expo-router": "^2.0.0",
     "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
@@ -45,6 +47,7 @@
     "react-native": "0.72.6",
     "react-native-animatable": "^1.4.0",
     "react-native-dotenv": "^3.4.9",
+    "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "~2.12.0",
     "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
@@ -52,6 +55,7 @@
     "react-native-screens": "~3.22.0",
     "react-native-tailwindcss": "^1.1.11",
     "react-native-url-polyfill": "^2.0.0",
+    "react-native-view-shot": "^3.8.0",
     "react-native-web": "~0.19.6",
     "sqlite-vss": "^0.1.2",
     "tailwind-rn": "^4.2.0",
